apps/api/apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.3
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
httpx==0.27.2
tenacity==9.0.0
